We will switch BufferMgr for the other strategies, so we will disassemble the test setup code.
